Iteration7: Protect book Routes with Authentication

+++ controllers/bookControllers.js
@@ -12,14 +12,30 @@ const getAllBooks = async (req, res) => {
 };
 
 // POST /books
+// const createBook = async (req, res) => {
+//   try {
+//     const newBook = await Book.create({ ...req.body});
+//     res.status(201).json(newBook);
+//   }catch (error) {
+//     res.status(400).json({message: "Failed to create book", error: error.message});
+//   }
+//   // res.send("createBook");
+// };
+
+// Create a new book
 const createBook = async (req, res) => {
   try {
-    const newBook = await Book.create({ ...req.body});
+    const user_id = req.user._id;
+    const newBook = new Book({
+      ...req.body,
+      user_id,
+    });
+    await newBook.save();
     res.status(201).json(newBook);
-  }catch (error) {
-    res.status(400).json({message: "Failed to create book", error: error.message});
+  } catch (error) {
+    console.error("Error creating book:", error);
+    res.status(500).json({ error: "Server Error" });
   }
-  // res.send("createBook");
 };
 
 // GET /books/:bookId
@@ -50,7 +66,9 @@ const updateBook = async (req, res) => {
   }
   try {
     const updatedBook = await Book.findOneAndUpdate(
-      {_id: bookId},
+      {_id: bookId,
+      user_id: req.user._id}, 
+
       {
         ...req.body
       },
@@ -78,7 +96,7 @@ const deleteBook = async (req, res) => {
   }
 
   try {
-    const deletedBook = await Book.findOneAndDelete({_id: bookId});
+    const deletedBook = await Book.findOneAndDelete({_id: bookId, user_id: req.user._id});
     if (deletedBook) {
       res.status(204).send();
